refactor: split test_run.py into multiple modules to facilitate addition of new tests in the future
test order is specified with the marker pytest.mark.order provided by pytest-order extension.
