+++ test/cli/__init__.py
@@ -1,0 +1,1 @@
+"""Package for testing the command line interface of DataHIPy."""
